Require either $content or $source to be set in nrpe::plugin.
If none of the two are set puppet will still create an empty executable
file which will result in the nrpe error message:
"NRPE: Unable to read output"

However the check will still appear as OK in nagios. This patch
explicitly checks that precisely one of the two options is set.